% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Odyssey Shiny Application Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
The Shiny server function for the Odyssey app.
}
\description{
Defines the server logic of the Odyssey Shiny app.
}
