% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Odyssey Shiny Application UI}
\usage{
app_ui(request)
}
\arguments{
\item{request}{The request object passed to the Shiny app.}
}
\value{
A Shiny UI definition.
}
\description{
Defines the user interface of the Odyssey Shiny app.
}
