% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{download_server}
\alias{download_server}
\title{Server Module: Table Tab - Download Dataset}
\usage{
download_server(id, df)
}
\arguments{
\item{id}{Character string for namespacing the module}

\item{df}{A reactive \code{data.table} containing the dataset to be downloaded}
}
\value{
A Shiny download handler that exports \code{data.table} as a CSV file
}
\description{
This server module provides a download handler for the dataset,
allowing users to export the current data as a CSV file.
}
