% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{downloadServer}
\alias{downloadServer}
\title{Table Tab: Download Dataset}
\usage{
downloadServer(id, df)
}
\arguments{
\item{id}{Character string for namespacing the module}

\item{df}{A reactive \code{data.table} containing the dataset to be downloaded}
}
\value{
A Shiny download handler that exports \code{data.table} as a CSV file
}
\description{
A Shiny server module that provides a download handler for the dataset,
allowing users to export the current data as a CSV file.
}
