% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_header.R
\name{header_ui}
\alias{header_ui}
\title{Table: Source Selection UI}
\usage{
header_ui(id)
}
\arguments{
\item{id}{Character string used for namespacing the input IDs in the UI module.}
}
\value{
A \code{tagList} with UI elements for selecting the data source and filters.
}
\description{
This UI module displays inputs for selecting a data source, country, date range,
and a button to trigger loading.
}
