% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pop_up.R
\name{info_ui}
\alias{info_ui}
\title{Display Instructions Modal for the App}
\usage{
info_ui()
}
\value{
A \code{shiny::modalDialog} object to be used with \code{shiny::showModal()}.
}
\description{
This function creates a reusable modal dialog in Shiny that provides users
with instructions on how to use the app. It includes a welcome message,
information about the data sources (ENA and GBIF), and step-by-step guidance
for navigating the app.
}
