% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map.R
\name{map_ui}
\alias{map_ui}
\title{UI Module: Map Tab}
\usage{
map_ui(id)
}
\arguments{
\item{id}{Character string used for namespacing the input IDs in the UI module.}
}
\value{
A \code{tagList} with UI elements for selecting the data source and filters.
}
\description{
Defines the user interface for the Map tab of the Odyssey Shiny application.
This tab displays an interactive Leaflet map that visualizes molecular biodiversity data.
}
