% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview.R
\name{plot_server3}
\alias{plot_server3}
\title{Overview Tab: Wordcloud of Scientific Names}
\usage{
plot_server3(id, df)
}
\arguments{
\item{id}{numeric identifier}

\item{df}{data table}
}
\description{
Overview Tab: Wordcloud of Scientific Names
}
