% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview.R
\name{plot_server4}
\alias{plot_server4}
\title{Server Module: Overview Tab - Pie Chart with the Isolation Source}
\usage{
plot_server4(id, df)
}
\arguments{
\item{id}{numeric identifier}

\item{df}{data table}
}
\description{
Server Module: Overview Tab - Pie Chart with the Isolation Source
}
