% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{plotServer2}
\alias{plotServer2}
\title{Overview Tab: Bar plot of Tax Divisions}
\usage{
plotServer2(id, df)
}
\arguments{
\item{id}{numeric identifier}

\item{df}{data table}
}
\description{
Overview Tab: Bar plot of Tax Divisions
}
