% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{plotServer3}
\alias{plotServer3}
\title{Overview Tab: Wordcloud of Scientific Names}
\usage{
plotServer3(id, df)
}
\arguments{
\item{id}{numeric identifier}

\item{df}{data table}
}
\description{
Overview Tab: Wordcloud of Scientific Names
}
