% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{plotServer4}
\alias{plotServer4}
\title{Overview Tab: Pie Chart with the Isolation Source}
\usage{
plotServer4(id, df)
}
\arguments{
\item{id}{numeric identifier}

\item{df}{data table}
}
\description{
Overview Tab: Pie Chart with the Isolation Source
}
