% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar.R
\name{source_ui}
\alias{source_ui}
\title{UI Module: Sidebar - Source Selection}
\usage{
source_ui(id)
}
\arguments{
\item{id}{Character string used for namespacing the input IDs in the UI module.}
}
\value{
A \code{tagList} with UI elements for selecting the data source and filters.
}
\description{
Generates the sidebar user interface for selecting data sources and filtering options
This module allows users to choose the input source,
select a country, define a date range, and trigger data loading.
}
