% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar.R
\name{table_options_ui}
\alias{table_options_ui}
\title{UI Module: Sidebar - Table Options}
\usage{
table_options_ui(id)
}
\arguments{
\item{id}{Character string used for namespacing the input IDs in the UI module.}
}
\value{
A \code{tagList} with UI elements for table customization.
}
\description{
This UI module displays checkboxes to show filters and group data by selected categories.
}
