% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{table_ui}
\alias{table_ui}
\title{UI Module: Table Tab - Table Display}
\usage{
table_ui(id)
}
\arguments{
\item{id}{Character string used for namespacing the input IDs in the UI module.}
}
\value{
A \code{nav_panel} containing the data table and download button.
}
\description{
This UI module defines the table panel layout for displaying data in a reactive table
and provides an option to download the results as a CSV file.
}
