% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{tableOptions}
\alias{tableOptions}
\title{Sidebar: Table Options UI}
\usage{
tableOptions(id)
}
\arguments{
\item{id}{Character string used for namespacing the input IDs in the UI module.}
}
\value{
A \code{tagList} with UI elements for table customization.
}
\description{
This UI module displays checkboxes to show filters and group data by selected categories.
}
