% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{text_server1}
\alias{text_server1}
\title{Server Module: Table Tab - Row Count Display}
\usage{
text_server1(id, df)
}
\arguments{
\item{id}{Character string for namespacing the module}

\item{df}{A reactive \code{data.table}; the dataset for which the number of observations is computed.}
}
\value{
A character string with the formatted number of rows
}
\description{
This server module returns a formatted text output displaying
the number of rows (observations) in the reactive dataset.
It is typically used in the Table tab to show the total count of records.
}
