% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{text_server2}
\alias{text_server2}
\title{Server Module: Table Tab - Number of Unique Taxonomic Divisions}
\usage{
text_server2(id, df)
}
\arguments{
\item{id}{Character string for namespacing the module}

\item{df}{A reactive \code{data.table}; the dataset for which unique taxonomic divisions are counted}
}
\value{
A numeric value representing the count of unique \code{tax_division2} entries
}
\description{
This server module returns the number of unique taxonomic divisions
in the dataset, typically used in the Table Tab to summarize diversity.
}
