% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{text_server3}
\alias{text_server3}
\title{Table Tab: Number of Unique Scientific Names}
\usage{
text_server3(id, df)
}
\arguments{
\item{id}{Character string for namespacing the module}

\item{df}{A reactive \code{data.table}; the dataset for which unique scientific names are counted}
}
\value{
A numeric value representing the count of unique \code{scientific_name} entries
}
\description{
A Shiny server module that returns the number of unique scientific names
in the dataset, typically used in the Table Tab to summarize species diversity.
}
