% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataset.R
\name{data_server}
\alias{data_server}
\title{ENA Query}
\usage{
data_server(id)
}
\arguments{
\item{id}{Character string for namespacing the module (used with \code{moduleServer}).}
}
\value{
A reactive expression that returns a \code{data.table} of the downloaded data.
}
\description{
This Shiny server module retrieves sequence data from the ENA database
based on user input for country and date range.
}
