% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{datasetServer}
\alias{datasetServer}
\title{Table Tab}
\usage{
datasetServer(id, df)
}
\arguments{
\item{id}{numeric identifier}

\item{df}{A reactive expression returning a data.frame containing the dataset.}
}
\description{
Table Tab
}
