% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home.R
\name{home_server}
\alias{home_server}
\title{Server Module: Home Tab}
\usage{
home_server(id)
}
\arguments{
\item{id}{Character string specifying the module namespace identifier.}
}
\value{
This function is called for its side effects to render UI output.
}
\description{
Defines the server logic for the Home tab of the Odyssey Shiny application.
This tab includes a welcome message, project description, intended audience, methodology,
contribution guidelines, and licensing information.
}
