% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home.R
\name{home_ui}
\alias{home_ui}
\title{UI Module: Home tab}
\usage{
home_ui(id)
}
\arguments{
\item{id}{A character string used to specify the module namespace.}
}
\value{
A Shiny \code{nav_panel} containing the "Home" section layout.
}
\description{
A Shiny UI module that defines the "Home" tab of the Odyssey app interface.
It displays introductory content, such as app information and
welcome text.
}
