% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_map.R
\name{map_server}
\alias{map_server}
\title{Map tab}
\usage{
map_server(id, df)
}
\arguments{
\item{id}{Character string specifying the module namespace identifier.}

\item{df}{A reactive \code{data.table} containing sequence records.}
}
\value{
A \code{leaflet} map rendered in the UI.
}
\description{
A Shiny server module that renders an interactive leaflet map showing
the locations of samples. Points are clustered and include
popups with metadata (accession, taxonomic division, scientific name).
}
