% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview.R
\name{plot_server2}
\alias{plot_server2}
\title{Overview Tab: Bar plot of Tax Divisions}
\usage{
plot_server2(id, df)
}
\arguments{
\item{id}{numeric identifier}

\item{df}{data table}
}
\description{
Overview Tab: Bar plot of Tax Divisions
}
