% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_overview.R
\name{plot_server3}
\alias{plot_server3}
\title{Server Module: Overview Tab - Wordcloud of Scientific Names}
\usage{
plot_server3(id, df)
}
\arguments{
\item{id}{numeric identifier}

\item{df}{data table}
}
\description{
Server Module: Overview Tab - Wordcloud of Scientific Names
}
