% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{text_server1}
\alias{text_server1}
\title{Table Tab: Number of Rows}
\usage{
text_server1(id, df)
}
\arguments{
\item{id}{Character string for namespacing the module}

\item{df}{A reactive \code{data.table}; the dataset for which the number of observations is computed.}
}
\value{
A character string with the formatted number of rows
}
\description{
A Shiny server module that returns the number of rows in the dataset.
Typically used in the Table Tab to display the total number of observations.
}
